refactor: extract whisplay gpiod shim module
